Update skywatcher.c - park rotator, improved rotator movement
Added function to park the rotator.

Changed behavior for incoming requests to change position. If the motor is moving, the requests are silently ignored. Stopping the motor movement before each request to change position caused the movement from point A to point B to take a very long time because tracking applications send the dynamic position of the tracked object (e.g. satellite) in regular short intervals and in case the rotator was on the opposite side, by the time the rotator reached the position of the object, the object had traveled a considerable distance.