Yaesu FTX1 Changes
When I tried to connect rigctld to wsjtx for this rig I could get freq but not set.  Had to adjust the IF response for 30.  Then it would cause weird hangs which was resolved by increasing the FTX1_POST_WRITE_DELAY from 2 to 5.  Now seems to work fine with initial testing in WSJT and rigctld.  Also the band mapping was a little different for 2m 70cm.