First cut at MacOS workflows with tar'd artifacts
Unfortunately, unlike with Ubuntu, MacOS is not available in act, so
these are not tested until they get pushed to master on GitHub.

The objective of these files is to generate a tarball of all build
artifacts so that a build failure can be more fully investigated.  I ran
out of time to figure out how to come up with a way to have unique tar
file names.  Separate files will do for now.