Advance master branch to 5.0.0~git
Reference GitHub issue #1773.  The forthcoming release will have API and
ABI changes and the ABI version is bumped to reflect that there is no
binary compatibility with prior releases.

It is planned that the 4.7.x series will be supported for some time,
possibly in parallel for a time with 5.0.x until downstream apps can
update.