FTX-1: Declare MAIN/SUB VFO ranges for correct vfo_fixup behavior
Change rx/tx range lists from RIG_VFO_ALL to RIG_VFO_MAIN | RIG_VFO_SUB.
This might be more accurate for the FTX-1 since it is a MAIN/SUB radio
and causes VFO_HAS_MAIN_SUB_ONLY to be true. This ensures vfo_fixup()
preserves MAIN as MAIN instead of converting to VFOA which shares cache
slots and might be part (not all) of the reason GPredict was having trouble.